+++ src/test_matpack.cc
@@ -55,7 +55,7 @@ int test1()
   cout << "v.nelem() = " << v.nelem() << "\n";
 
   for (Index i=0; i<v.nelem(); ++i )
-    v[i] = i;
+    v[i] = (Numeric)i;
 
   cout << "v.begin() = " << *v.begin() << "\n";
 
@@ -71,7 +71,7 @@ int test1()
   for (Index i=0 ; i<1000; ++i)
     {
       Vector v3(1000);
-      v3 = i;
+      v3 = (Numeric)i;
     }
 
   v2[Range(joker)] = 88;
@@ -174,8 +174,8 @@ void test4()
   
   for ( Index i=0; i<a.nelem(); ++i )
     {
-      a[i] = i+1;
-      b[i] = a.nelem()-i;
+      a[i] = (Numeric)i+1;
+      b[i] = (Numeric)(a.nelem()-i);
     }
 
   cout << "a = \n" << a << "\n";
@@ -252,7 +252,7 @@ void test8()
 {
   Vector x(80000000);
   for ( Index i=0; i<x.nelem(); ++i )
-    x[i] = i;
+    x[i] = (Numeric)i;
   cout << "Done." << "\n";
 }
 
@@ -525,7 +525,7 @@ void test31()
   for ( Index i=0; i<a.npages(); ++i )
     for ( Index j=0; j<a.nrows(); ++j )
       for ( Index k=0; k<a.ncols(); ++k )
-        a(i,j,k) = ++fill;
+        a(i,j,k) = (Numeric)(++fill);
 
   cout << "a =\n" << a << "\n";
 
@@ -559,7 +559,7 @@ void test31()
 
   Index s = 200;
   cout << "Let's allocate a large tensor, "
-       << s*s*s*8/1024./1024.
+       << (Numeric)(s*s*s)*8/1024./1024.
        << " MB...\n";
 
   a.resize(s,s,s);
@@ -576,7 +576,7 @@ void test31()
   for ( Index i=0; i<a.npages(); ++i )
     for ( Index j=0; j<a.nrows(); ++j )
       for ( Index k=0; k<a.ncols(); ++k )
-        a(i,j,k) = ++fill;
+        a(i,j,k) = (Numeric)(++fill);
 
   cout << "Max(a) = ...\n";
